// frontend/src/app/shared/generated-api/model/authenticatedUser.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthenticatedUser { 
    id: string;
    username: string;
    email: string;
    displayName?: string;
    role: AuthenticatedUserRoleEnum;
}
export enum AuthenticatedUserRoleEnum {
    User = 'USER',
    Admin = 'ADMIN'
};