// frontend/src/app/shared/generated-api/model/loginOutcome.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MfaChallenge } from './mfaChallenge';
import { SessionContext } from './sessionContext';


export interface LoginOutcome { 
    result: LoginOutcomeResultEnum;
    context?: SessionContext;
    challenge?: MfaChallenge;
}
export enum LoginOutcomeResultEnum {
    Session = 'SESSION',
    MfaRequired = 'MFA_REQUIRED'
};