// frontend/src/app/shared/generated-api/model/loginRequest.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginRequest { 
    identifier: string;
    password: string;
    rememberMe: boolean;
}